minor fixes

+++ ft_split.c
@@ -6,7 +6,7 @@
 /*   By: mvan-der <mvan-der@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/11/07 16:18:32 by mvan-der      #+#    #+#                 */
-/*   Updated: 2020/11/26 14:56:17 by mvan-der      ########   odam.nl         */
+/*   Updated: 2020/11/26 15:03:25 by mvan-der      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 static char	**ft_free(char **result)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (result[i])
@@ -29,8 +29,8 @@ static char	**ft_free(char **result)
 
 static int	ft_cnt_str(const char *s, char c)
 {
-	int		i;
-	int		count;
+	size_t	i;
+	size_t	count;
 	int		checkflag;
 
 	i = 0;
@@ -52,7 +52,7 @@ static int	ft_cnt_str(const char *s, char c)
 
 static int	ft_len_str(const char *s, int i, char c)
 {
-	int len;
+	size_t len;
 
 	len = 0;
 	while (s[i] && s[i] != c)
@@ -65,7 +65,7 @@ static int	ft_len_str(const char *s, int i, char c)
 
 static void	ft_split_magic(char const *s, char c, char **result)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (s[i])
